Time: 0 ms (100.00%) | Memory: 8.5 MB (31.81%) - LeetSync

+++ 3275-minimum-number-of-pushes-to-type-word-i/minimum-number-of-pushes-to-type-word-i.cpp
@@ -3,7 +3,7 @@ public:
     int minimumPushes(string word) {
         vector<int> v(26,0);
         for(auto& c: word){
-            v[(c- 'a')]++;
+            v[static_cast<int>(c- 'a')]++;
         }
         sort(v.begin(), v.end());
